+++ HISTORY.md
@@ -7,6 +7,8 @@ Dev
 
 * Fixes a bug that report *page-count* as invalid when fpage or lpage values 
   are non-digit.
+* Fixes a bug that mark as invalid XMLs containing use-licenses with 
+  https scheme or missing trailing slashes.
 
 
 0.7 (2015-03-13)